+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `laravel-captcha` will be documented in this file.
+
+## 1.0.0 - 2026-03-21
+
+- Initial release
+- Support for Google reCAPTCHA (v2 & v3)
+- Support for Cloudflare Turnstile
+- Support for hCaptcha
+- Validation rule (`CaptchaRule`)
+- Facade support
+- Global enable/disable configuration